refactor(gym): split tests_scenarios.rs (#1266)
tests_scenarios.rs 1183 -> 398 LOC; new _2/_3/_4/_5 (~320/378/306/273).
