+++ src/gym/mod.rs
@@ -23,6 +23,14 @@ mod tests_reporting_more;
 #[cfg(test)]
 mod tests_scenarios;
 #[cfg(test)]
+mod tests_scenarios_2;
+#[cfg(test)]
+mod tests_scenarios_3;
+#[cfg(test)]
+mod tests_scenarios_4;
+#[cfg(test)]
+mod tests_scenarios_5;
+#[cfg(test)]
 mod tests_types;
 #[cfg(test)]
 mod tests_types_extra;
