+++ src/gym/tests_scenarios.rs
@@ -395,789 +395,3 @@ fn class_checks_test_writing_passes_with_keywords() {
             .any(|c| c.id == "test-covers-basic-case" && c.passed)
     );
 }
-
-#[test]
-fn class_checks_test_writing_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::TestWriting,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-// -- SessionQuality checks --
-
-#[test]
-fn class_checks_session_quality_passes_with_nonempty_summary_and_enough_memory() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::SessionQuality,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("plan", "non-empty execution summary", "reflection");
-    let exported = dummy_handoff(3);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 1);
-    assert!(checks[0].passed);
-    assert_eq!(checks[0].id, "session-quality-summary-adequate");
-    assert!(checks[0].detail.contains("3 memory records"));
-    assert!(checks[0].detail.contains("non-empty"));
-}
-
-#[test]
-fn class_checks_session_quality_fails_with_empty_summary() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::SessionQuality,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("plan", "   ", "reflection");
-    let exported = dummy_handoff(5);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 1);
-    assert!(!checks[0].passed);
-    assert!(checks[0].detail.contains("empty"));
-}
-
-#[test]
-fn class_checks_session_quality_fails_with_insufficient_memory() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::SessionQuality,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("plan", "has content", "reflection");
-    let exported = dummy_handoff(1);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 1);
-    assert!(!checks[0].passed);
-}
-
-// -- Edge cases for keyword detection --
-
-#[test]
-fn class_checks_repo_exploration_detects_module_keyword() {
-    let scenario = repo_exploration_scenario();
-    let outcome = dummy_outcome("found module layout", "bland", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "repo-structure-discovered")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_repo_exploration_detects_crate_for_deps() {
-    let scenario = repo_exploration_scenario();
-    let outcome = dummy_outcome("bland", "uses crate dependencies", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "repo-dependencies-identified")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_repo_exploration_detects_lib_rs_for_entry_points() {
-    let scenario = repo_exploration_scenario();
-    let outcome = dummy_outcome("bland", "bland", "found lib.rs");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "repo-entry-points-found")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_documentation_detects_doc_comment_keyword() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::Documentation,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("doc comment style", "bland", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "doc-comment-syntax-valid")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_safe_code_change_detects_cargo_build() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::SafeCodeChange,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("bland", "ran cargo build successfully", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "code-change-compilation-checked")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_test_writing_detects_expect_for_assertions() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::TestWriting,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("bland", "test uses expect to verify", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "test-has-assertions")
-        .unwrap();
-    assert!(check.passed);
-}
-
-// -- BugFix checks --
-
-#[test]
-fn class_checks_bug_fix_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::BugFix,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "found unwrap bug in error handling path",
-        "fix: replace expect with Result propagation for safety",
-        "the panic was unsafe, now uses graceful error recovery",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "bug-defect-identified" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "bug-fix-described" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "bug-safety-analyzed" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_bug_fix_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::BugFix,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_bug_fix_detects_panic_keyword() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::BugFix,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("found a panic in production code", "bland", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "bug-defect-identified")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_bug_fix_detects_convert_for_fix() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::BugFix,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("bland", "convert the call to use ? operator", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks.iter().find(|c| c.id == "bug-fix-described").unwrap();
-    assert!(check.passed);
-}
-
-// -- Refactoring checks --
-
-#[test]
-fn class_checks_refactoring_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::Refactoring,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "extract helper fn from long function",
-        "refactored code preserves original behavior",
-        "before and after code shown, equivalent output",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "refactor-change-identified" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "refactor-behavior-preserved" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "refactor-code-shown" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_refactoring_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::Refactoring,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_refactoring_detects_simplify_keyword() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::Refactoring,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("simplified the match expression", "bland", "bland");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "refactor-change-identified")
-        .unwrap();
-    assert!(check.passed);
-}
-
-#[test]
-fn class_checks_refactoring_detects_preserve_keyword() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::Refactoring,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("bland", "bland", "behavior is preserved after change");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    let check = checks
-        .iter()
-        .find(|c| c.id == "refactor-behavior-preserved")
-        .unwrap();
-    assert!(check.passed);
-}
-
-// -- Coverage distribution checks --
-
-#[test]
-fn benchmark_scenarios_each_class_has_at_least_two() {
-    let scenarios = benchmark_scenarios();
-    let count = |class: BenchmarkClass| scenarios.iter().filter(|s| s.class == class).count();
-    assert!(count(BenchmarkClass::RepoExploration) >= 2);
-    assert!(count(BenchmarkClass::Documentation) >= 2);
-    assert!(count(BenchmarkClass::SafeCodeChange) >= 2);
-    assert!(count(BenchmarkClass::SessionQuality) >= 2);
-    assert!(count(BenchmarkClass::TestWriting) >= 2);
-    assert!(count(BenchmarkClass::BugFix) >= 2);
-    assert!(count(BenchmarkClass::Refactoring) >= 2);
-    assert!(count(BenchmarkClass::DependencyAnalysis) >= 2);
-    assert!(count(BenchmarkClass::ErrorHandling) >= 2);
-    assert!(count(BenchmarkClass::DataMigration) >= 2);
-    assert!(count(BenchmarkClass::CicdPipeline) >= 2);
-    assert!(count(BenchmarkClass::DependencyUpgrade) >= 2);
-    assert!(count(BenchmarkClass::ReleaseManagement) >= 2);
-    assert!(count(BenchmarkClass::AccessibilityReview) >= 2);
-    assert!(count(BenchmarkClass::InternationalizationReview) >= 2);
-    assert!(count(BenchmarkClass::IncidentResponse) >= 2);
-}
-
-#[test]
-fn benchmark_scenarios_multi_process_coverage() {
-    let multi = benchmark_scenarios()
-        .iter()
-        .filter(|s| s.topology == RuntimeTopology::MultiProcess)
-        .count();
-    assert!(
-        multi >= 2,
-        "need at least 2 MultiProcess scenarios, got {multi}"
-    );
-}
-
-#[test]
-fn benchmark_scenarios_copilot_sdk_coverage() {
-    let copilot = benchmark_scenarios()
-        .iter()
-        .filter(|s| s.base_type == "copilot-sdk")
-        .count();
-    assert!(
-        copilot >= 2,
-        "need at least 2 copilot-sdk scenarios, got {copilot}"
-    );
-}
-
-// -- DependencyAnalysis checks --
-
-#[test]
-fn class_checks_dependency_analysis_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DependencyAnalysis,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "analyzed cargo.toml dependencies and crate versions",
-        "import coupling between modules assessed",
-        "suggest decoupling the most coupled module",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-analysis-performed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-coupling-assessed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-recommendations-present" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_dependency_analysis_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DependencyAnalysis,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-// -- ErrorHandling checks --
-
-#[test]
-fn class_checks_error_handling_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::ErrorHandling,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "found unwrap calls that could panic in production",
-        "classified as safe or risky based on context",
-        "error propagation chain traced through diagnostic output",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "error-analysis-performed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "error-classification-present" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "error-propagation-traced" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_error_handling_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::ErrorHandling,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-// -- Distributed topology coverage --
-
-#[test]
-fn benchmark_scenarios_distributed_topology_coverage() {
-    let distributed = benchmark_scenarios()
-        .iter()
-        .filter(|s| s.topology == RuntimeTopology::Distributed)
-        .count();
-    assert!(
-        distributed >= 1,
-        "need at least 1 Distributed scenario, got {distributed}"
-    );
-}
-
-// -- Wave 7: DataMigration / CicdPipeline / DependencyUpgrade / ReleaseManagement --
-
-#[test]
-fn class_checks_data_migration_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DataMigration,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "schema delta adds optional field with serde default",
-        "backward compatibility preserved during phased rollout",
-        "rollback path documented if migration must be reverted",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "data-migration-schema-delta-described" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "data-migration-compatibility-addressed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "data-migration-rollout-or-rollback-planned" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_data_migration_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DataMigration,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_cicd_pipeline_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::CicdPipeline,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "drafted github actions workflow .yml with build job and steps",
-        "trigger on pull_request and push, pin uses: actions/checkout@v4",
-        "runs cargo check and cargo test with cache and matrix",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "cicd-workflow-structure-described" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "cicd-trigger-or-pin-addressed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "cicd-verification-or-remediation-present" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_cicd_pipeline_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::CicdPipeline,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_dependency_upgrade_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DependencyUpgrade,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "plan major version bump of crate in cargo.toml",
-        "changelog lists breaking api change at call site",
-        "verify with cargo check and cargo test, staged rollout with rollback",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-upgrade-target-named" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-upgrade-breakage-analyzed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "dep-upgrade-verification-plan-present" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_dependency_upgrade_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::DependencyUpgrade,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_release_management_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::ReleaseManagement,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "semver minor version bump in cargo.toml",
-        "changelog grouped by added/changed/fixed for release notes",
-        "git tag and publish, with rollback path",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "release-version-bump-planned" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "release-changelog-authored" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "release-tag-or-cutover-addressed" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_release_management_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::ReleaseManagement,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-// -- Wave 8: AccessibilityReview / InternationalizationReview / IncidentResponse --
-
-#[test]
-fn class_checks_accessibility_review_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::AccessibilityReview,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "audit aria roles, missing alt text, label association, and keyboard focus",
-        "cite WCAG 2.1.1 and 4.1.2 success criterion at level AA",
-        "remediation: add aria-label, fix focus order, replace contrast pair",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "a11y-issues-identified" && c.passed)
-    );
-    assert!(checks.iter().any(|c| c.id == "a11y-wcag-cited" && c.passed));
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "a11y-remediation-proposed" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_accessibility_review_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::AccessibilityReview,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_internationalization_review_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::InternationalizationReview,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "inventory hardcoded string literals and route to message catalog",
-        "negotiate locale via Accept-Language with CLDR fallback (en-US, pt-BR)",
-        "address plural categories, RTL/bidi mirroring, and date format via ICU MessageFormat",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "i18n-localizable-strings-identified" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "i18n-locale-handling-described" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "i18n-pluralization-or-format-addressed" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_internationalization_review_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::InternationalizationReview,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
-
-#[test]
-fn class_checks_incident_response_passes_with_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::IncidentResponse,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome(
-        "reconstruct timeline: alert paged, mitigation started, resolved at",
-        "blameless analysis: root cause distinct from trigger, latent contributing factor",
-        "follow-up runbook with on-call escalation and prevention action item",
-    );
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "incident-timeline-reconstructed" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "incident-root-cause-or-contributing-identified" && c.passed)
-    );
-    assert!(
-        checks
-            .iter()
-            .any(|c| c.id == "incident-mitigation-or-followup-proposed" && c.passed)
-    );
-}
-
-#[test]
-fn class_checks_incident_response_fails_without_keywords() {
-    let scenario = BenchmarkScenario {
-        class: BenchmarkClass::IncidentResponse,
-        ..repo_exploration_scenario()
-    };
-    let outcome = dummy_outcome("nothing", "bland text", "empty");
-    let exported = dummy_handoff(0);
-    let checks = class_specific_checks(&scenario, &outcome, &exported);
-    assert_eq!(checks.len(), 3);
-    for check in &checks {
-        assert!(!check.passed, "check '{}' should have failed", check.id);
-    }
-}
